New Agents/requirements.txt
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
logging>=0.5.1.2 